refactor: move node actions to dropdown
